% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhatt_dist.R
\name{bhatt_dist}
\alias{bhatt_dist}
\title{Bhattacharyya distance}
\usage{
bhatt_dist(d1, d2, ...)
}
\arguments{
\item{d1}{matrix or numeric dataframe}

\item{d2}{matrix or numeric dataframe}

\item{...}{additional arguments to be passed to \code{cov} or \code{mean}}
}
\value{
Bhattacharyya distance
}
\description{
Measures the distance between two multivariate distributions
}
\examples{
bhatt_dist(iris[,1:4], iris[,1:4])
bhatt_dist(iris[1:10, 1:4], iris[20:30, 1:4])

}

