% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centre_control.R
\name{centre_control}
\alias{centre_control}
\title{Shift co-ordinates to centre on a compounds centroid}
\usage{
centre_control(df, x, y, cmpd_col, cmpd)
}
\arguments{
\item{df}{dataframe}

\item{x}{column of x co-ordinates}

\item{y}{column of y-co-ordinates}

\item{cmpd_col}{column containing names to subset with \code{cmpd}}

\item{cmpd}{name of compound to subset \code{cmpd_col} with}
}
\value{
dataframe containing x, y, cmpd_col columns shifted so that
  cmpd is centered on 0,0
}
\description{
Idea is to shift co-ordinates so that the centroid of the negative control
is always centered on 0,0. Useful for PCA biplots where the 0,0 denotes the
centroid of the entire data, and not necessarily the negative control. If 
given a dataframe and specified columns for the x and y co-ordinates, this
function will shift the co-ordinates to be centered on the centroid of a
specified compound.
}
\examples{
# artifical example using iris
pca <- prcomp(iris[,1:4])$x[,1:2]
df <- data.frame(pca, name = iris[,5])
pca_shift <- centre_control(df, x = 'PC1', y = 'PC2', cmpd_col = 'name', cmpd = 'setosa')

par(mfrow = c(2,1))
plot(df$PC1, df$PC2, col = df$name, pch = 20)
plot(pca_shift$PC1, pca_shift$PC2, col = pca_shift$name, pch = 20)
}

