% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_pairs.R
\name{cosine_pairs}
\alias{cosine_pairs}
\title{Calculate pairs of cosine similarities between replicates}
\usage{
cosine_pairs(x, cols)
}
\arguments{
\item{x}{list}

\item{cols}{integer, column indices}
}
\value{
dataframe of compound combinations across replicates with a column 
   of cosine similarity values
}
\description{
Given a list, with an element per compound, \code{cols}
are the elements of the vector with which to calculate the cosine
similarity
}
\examples{
cmpds <- c(rep('a', 100), rep('b', 100), rep('c', 100))
replicate <- rep(1:100, 3)
PC1 <- rnorm(300)
PC2 <- rnorm(300)

df <- data.frame(cmpds, replicate, PC1, PC2)

df_split <- split(df, df$cmpds)

# works with unequal replicate sizes
df_split$a <- df_split$a[-c(1:10), ]

cosine_pairs(df_split, 3:4)
}

