% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sim.R
\name{cosine_sim_vector}
\alias{cosine_sim_vector}
\title{Cosine similarity between two vectors}
\usage{
cosine_sim_vector(a, b, ...)
}
\arguments{
\item{a}{vector}

\item{b}{vector}

\item{...}{optional arguments to sum (i.e na.rm = TRUE)}
}
\value{
out cosine similarity
}
\description{
Given two vectors, this function will calculate a cosine similarity.
Used as an internal function for \code{cosine_sim()}
}
\examples{
set.seed(12321)
x <- rnorm(20)
y <- rnorm(20)
cosine_sim_vector(x, y)
}

