\name{distToNeg}
\alias{distToNeg}
\title{ distToNeg }
\description{
  From a pair-wise distance matrix, this function extracts the corresponding distance between a sample and negative controls.
}
\usage{
distToNeg(x, distMatrix, neg='rluc')
}
\arguments{
  \item{x}{An \code{imageHTS} object }
  \item{distMatrix}{A pair-wise distance matrix, as generated by \code{PDMBySvmAccuracy}.}
  \item{neg}{A character string to identify negative controls.}
}
\details{
  
  This function averages the distance measurements between the sample and all negative controls on the same plate.
  
}
\value{
  A numeric vector with length equal to the dimension of the distance matrix.
}

\seealso{\code{PDMBySvmAccuracy}}

\author{Xian Zhang}

\examples{

  library('phenoDist')

  ## load the imageHTS object
  load(system.file('kimorph', 'kimorph.rda', package='phenoDist'))
  x@localPath <- file.path(tempdir(), 'kimorph')

  ## load sample phenotypic distance matrix
  load(system.file('kimorph', 'svmAccPDM_Pl1.rda', package='phenoDist'))

  ## replicate ranking
  ranking <- repDistRank(x, distMatrix=svmAccPDM_Pl1)
  summary(ranking)

  ## phenotypic distance to negative control
  pheno <- distToNeg(x, distMatrix=svmAccPDM_Pl1, neg='rluc')

  ## separation between negative and positive controls
  ctlSeparatn(x, pheno, neg='rluc', pos='ubc', method='robust')

  ## replicate correlation coefficient
  repCorr(x, pheno)
}
