% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid_dist.R
\name{euclid_dist}
\alias{euclid_dist}
\title{Euclidean distance between two vectors}
\usage{
euclid_dist(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
As \code{dist} in base R is idiotic.
}
\examples{
a <- rnorm(100)
b <- rnorm(100)
euclid_dist(a, b)
}

