% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_pairs.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{exapand grid, no equals}
\usage{
expand_grid_unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{include.equals}{whether to include duplicates, if TRUE will behave
just like expand.grid with two vectors}
}
\value{
dataframe
}
\description{
Like expand.grid but does not list repeats
}

