% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan distance between two vectors}
\usage{
manhattan(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric distance metric
}
\description{
A Manhattan or cityblock distance is the sum of the absolute difference
between all co-ordinates.
}
\examples{
a <- c(1,2,3,4,5,1,2,2)
b <- c(1,3,5,3,5,2,4,1)
manhattan(a, b)

}

