% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_z.R
\name{multi_z}
\alias{multi_z}
\title{Multivariate Z-prime}
\usage{
multi_z(df, feature_cols, cmpd_col, pos, neg, ...)
}
\arguments{
\item{df}{dataframe}

\item{feature_cols}{indices of the feature columns within \code{df}}

\item{cmpd_col}{index or column name of column within df that contains
the positive and negative control}

\item{pos}{name of positive control within cmpd_col}

\item{neg}{name of negative control within cmpd_col}

\item{...}{additional parameters is to be passed to \code{lda()}}
}
\value{
z-prime
}
\description{
Multivarite z-score as decribed by Kummel et al. Calculates a Z-prime
using a LDA projection of the multivariate data for the positive and
negative control.
}

