% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{r_zscore}
\alias{r_zscore}
\title{Robust Z-score}
\usage{
r_zscore(x, ...)
}
\arguments{
\item{x}{vector of numbers}

\item{...}{additional arguments for median and mad}
}
\description{
Robust version of the Z-score using the median and MAD rather than the
mean and standard deviation of the original metric. It is more resiliant
to outliers.
}

