% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmd.R
\name{ssmd_effect_message}
\alias{ssmd_effect_message}
\title{ssmd effect size description}
\usage{
ssmd_effect_message(beta)
}
\arguments{
\item{beta}{beta value from within \code{ssmd}}
}
\description{
Returns a ELI5 of the beta value
}

