% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta0.R
\name{theta0_}
\alias{theta0_}
\title{Angle between the vector and the origin}
\usage{
theta0_(a)
}
\arguments{
\item{a}{numeric vector}
}
\description{
Calculates and angle (in degrees) between the vector and the origin
\code{c(1, rep(0, length(vector) - 1))}
}

