% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_factor_scan.R
\name{z_factor_scan}
\alias{z_factor_scan}
\title{Multiple z-factor calculations}
\usage{
z_factor_scan(data, treatments, cutoff = 0.5, n = FALSE)
}
\arguments{
\item{data}{Dataframe, containing only numerical columns of features and
single column of treatment labels}

\item{treatments}{Vector of factors of treatment groups under the column of 'header'}

\item{cutoff}{Minimum z-factor threshold for returning featues. Default is 0.5}

\item{n}{Highest n features to be returned}
}
\description{
Calculates z-factors between two compounds for multiple variables and returns the
z-factors and associated features above a specified cut-off. The cut-off can be
determined either by a minumum z-factor score, or return the top n features.
}

