% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{zscore}
\alias{zscore}
\title{Z-score}
\usage{
zscore(x, ...)
}
\arguments{
\item{x}{vector of numbers}

\item{...}{additional arguments for mean and sd}
}
\description{
Scale and center data, essentially the same as \code{scale()}
}

